requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
coreapi==2.3.3
coreschema==0.0.4
Django==5.2.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
itypes==1.2.0
Jinja2==3.1.6
joblib==1.5.1
MarkupSafe==3.0.2
mysqlclient==2.2.7
numpy==2.3.0
PyJWT==2.9.0
PyMySQL==1.1.1
PyYAML==6.0.2
requests==2.32.3
scikit-learn==1.7.0
scipy==1.15.3
sqlparse==0.5.3
threadpoolctl==3.6.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
